AHYPRJ-tetechJournal/agent_social_media/requirements.txt
crewai>=0.11.0
# diffusers==0.32.2
# evolutionapi==0.1.1
Flask==3.1.0
# huggingface_hub==0.26.1
imgurpython==1.1.7
numpy==1.25.0
openai==1.76.2
pilgram==1.2.1
Pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
Requests==2.32.3
# torch==2.6.0
# transformers==4.48.3
fastapi
uvicorn
pydantic
python-telegram-bot==21.10
